feat(营销中心): 优惠券 1.用户优惠券列表增加状态查询条件

+++ src/main/java/com/acme/acmemall/model/OrderVo.java
@@ -161,7 +161,7 @@ public class OrderVo implements Serializable {
     @Builder.Default
     private BigDecimal freight_price = BigDecimal.ZERO;
     //使用的优惠券id
-    private Integer coupon_id;
+    private Long coupon_id;
     //
     private Integer parent_id;
     //优惠价格
